push

+++ backend/data/views.py
@@ -1,4 +1,6 @@
 #보안
+import os
+import jwt
 from dotenv import load_dotenv
 
 # API 제작 - 상태, 기능, 인증, 권한, Header(JSON)
@@ -39,29 +41,38 @@ def user_list(request):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 # User 하나만 보여줌
-@api_view(["GET", "PATCH", "DELETE"])
+@api_view(["GET", "PUT", "DELETE"])
 @parser_classes([JSONParser])
 @authentication_classes([SessionAuthentication, BasicAuthentication])
-def user_one(request, pk):
-    try:
-        user = User.objects.get(pk=pk)
-    except User.DoesNotExist():
+def user_one(request):
+    if not request.COOKIES.get("access_token"):
         return Response(status=status.HTTP_404_NOT_FOUND)
+    else:
+        try:
+            load_dotenv(verbose=True)
+            SECRET_KEY=os.getenv("SECRET_KEY")
+            ALGORITHM=os.getenv("ALGORITHM")
+            token=request.COOKIES.get("access_token")
+            payload=jwt.decode(token,SECRET_KEY,ALGORITHM)
+            user=User.objects.get(uid=payload["id"])
 
-    if request.method == "GET":
-        user_serializer = UserAllSerializer(user)
-        return Response(user_serializer.data)
-    
-    elif request.method == "PATCH":
-        user_serializer = UserAllSerializer(user, data=request.data)
-        if user_serializer.is_valid():
-            user_serializer.save()
+        except User.DoesNotExist():
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if request.method == "GET":
+            user_serializer = UserAllSerializer(user)
             return Response(user_serializer.data)
-        return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    elif request.method == 'DELETE':
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        
+        elif request.method == "PUT":
+            user_serializer = UserAllSerializer(user, data=request.data)
+            if user_serializer.is_valid():
+                user_serializer.save()
+                return Response(user_serializer.data)
+            return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
+        elif request.method == 'DELETE':
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
 # ServiceReview
 @api_view(["GET", "POST"])
